refactor(acp): extract provider switcher
